# api/openapi.yaml
openapi: 3.0.3
info:
  title: WalletHub API
  description: |
    WalletHub - платёжный шлюз для управления кошельками и транзакциями.

    ## Основные возможности
    - Управление пользователями и KYC верификация
    - Создание и управление мультивалютными кошельками (Fiat + Crypto)
    - Проведение финансовых операций (deposit, withdraw, transfer)
    - Полная история транзакций

    ## Аутентификация
    API использует Bearer Token аутентификацию. Передайте токен в заголовке:
    ```
    Authorization: Bearer <your-token>
    ```

    ## Идемпотентность
    Все финансовые операции поддерживают идемпотентность через `idempotency_key`.
    При повторном запросе с тем же ключом вернётся результат первой операции.

    ## Rate Limiting
    - Общий лимит: 100 запросов/минуту
    - Финансовые операции: 30 запросов/минуту

    Заголовки ответа:
    - `X-RateLimit-Limit`: Максимум запросов
    - `X-RateLimit-Remaining`: Оставшееся количество
    - `X-RateLimit-Reset`: Unix timestamp сброса

    ## Формат ответа
    Все ответы имеют единый формат:
    ```json
    {
      "success": true/false,
      "data": { ... },
      "error": { "code": "...", "message": "..." },
      "meta": { "page": 1, "total": 100 },
      "request_id": "uuid",
      "timestamp": "2024-01-01T00:00:00Z"
    }
    ```
  version: 1.0.0
  contact:
    name: WalletHub Support
    email: support@wallethub.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
...
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create user
      description: Create a new user with email and full name
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'

    get:
      tags: [Users]
      summary: List users
      description: Get paginated list of users
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Get user details by UUID
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/users/{id}/kyc:
    post:
      tags: [Users]
      summary: Approve/Reject KYC
      description: Approve or reject user's KYC verification
      operationId: approveKYC
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveKYCRequest'
      responses:
        '200':
          description: KYC status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/BusinessRuleError'

  # ============================================
  # Wallets
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BusinessRuleError:
      description: Business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ApiMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      required: [success, error, request_id, timestamp]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ApiError'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    ValidationErrorResponse:
      type: object
      required: [success, error, request_id, timestamp]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            fields:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
        details:
          type: object
    FieldError:
      type: object
      required: [field, message, code]
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string

    # ============================================
    # Health Schemas
    # ============================================
...
    CreateUserRequest:
      type: object
      required: [email, full_name]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        full_name:
          type: string
          minLength: 2
          maxLength: 100
          example: John Doe
    ApproveKYCRequest:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
        reason:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        kyc_status:
          $ref: '#/components/schemas/KYCStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    KYCStatus:
      type: string
      enum: [UNVERIFIED, PENDING, VERIFIED, REJECTED]
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    UserCreatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            message:
              type: string
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: '#/components/schemas/User'
            total_count:
              type: integer
        meta:
          $ref: '#/components/schemas/ApiMeta'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time

    # ============================================
    # Wallet Schemas
    # ============================================
...